+++ slack-service/src/main/java/com/gaebal_easy/client/slack/application/dto/SlackMessageInfoDTO.java
@@ -1,4 +1,4 @@
-package com.gaebal_easy.client.slack.presentation.dto;
+package com.gaebal_easy.client.slack.application.dto;
 
 import java.util.List;
 import java.util.UUID;
