+++ slack-service/src/main/java/com/gaebal_easy/client/slack/application/dto/SlackRequest.java
@@ -1,0 +1,5 @@
+package com.gaebal_easy.client.slack.application.dto;
+
+public class SlackRequest {
+
+}
