--- slack-service/src/main/java/com/gaebal_easy/client/slack/presentation/dto/SlackRequest.java
@@ -1,5 +1,0 @@
-package com.gaebal_easy.client.slack.presentation.dto;
-
-public class SlackRequest {
-
-}
